// src/preferencia/dto/update-preferencia.dto.ts
import { PartialType } from '@nestjs/mapped-types';
import { CreatePreferenciaDto } from './create-preferencia.dto';

export class UpdatePreferenciaDto extends PartialType(CreatePreferenciaDto) {
    PuedenIniciarSoloCompatibles?: string; 
    cantConversacionesMaximas?: number; 
    tieneNoMolestar?: boolean; 
}